+++ src/test/java/de/codescape/jira/plugins/scrumpoker/rest/SessionResourceTest.java
@@ -4,9 +4,9 @@ import com.atlassian.jira.security.JiraAuthenticationContext;
 import com.atlassian.jira.user.ApplicationUser;
 import de.codescape.jira.plugins.scrumpoker.ao.ScrumPokerSession;
 import de.codescape.jira.plugins.scrumpoker.rest.entities.SessionEntity;
+import de.codescape.jira.plugins.scrumpoker.service.EstimationFieldService;
 import de.codescape.jira.plugins.scrumpoker.service.ScrumPokerSessionService;
 import de.codescape.jira.plugins.scrumpoker.service.SessionEntityTransformer;
-import de.codescape.jira.plugins.scrumpoker.service.StoryPointFieldSupport;
 import org.junit.Rule;
 import org.junit.Test;
 import org.mockito.InjectMocks;
@@ -36,7 +36,7 @@ public class SessionResourceTest {
     private ScrumPokerSessionService scrumPokerSessionService;
 
     @Mock
-    private StoryPointFieldSupport storyPointFieldSupport;
+    private EstimationFieldService estimationFieldService;
 
     @Mock
     private SessionEntityTransformer sessionEntityTransformer;
@@ -101,7 +101,7 @@ public class SessionResourceTest {
         sessionResource.confirmEstimation(ISSUE_KEY, ESTIMATION);
 
         verify(scrumPokerSessionService, times(1)).confirm(ISSUE_KEY, USER_KEY, ESTIMATION);
-        verify(storyPointFieldSupport, times(1)).save(ISSUE_KEY, ESTIMATION);
+        verify(estimationFieldService, times(1)).save(ISSUE_KEY, ESTIMATION);
     }
 
     private void expectCurrentUserIs(String userKey) {